Digging Game 2: version 2.2.16 * set every planet's health to 500, reducing difficulty. (in "Sky Frame") * prevent double overflow by normalizing TILT. (in "Pod.cs") * use foreach to replace for. (in "Upgrade Plant.cs") * remove unnecessary try/catch block. (in "Block.cs") * try to dispose image data resource properly to reduce memory leak. (in "Game.cs") * change save format of money value from int to long to prevent overflow. (in "Statistics.cs") * detect missing directory and add show more error message. (in "Sound.cs")

+++ 鑽礦遊戲2/Game Frame/Statistics.cs
@@ -75,7 +75,7 @@ namespace 鑽礦遊戲2
             GasConsumed = double.Parse(data[idx++]);
             OresGet = int.Parse(data[idx++]);
             if (version <= 1) return;
-            MoneyCost = int.Parse(data[idx++]);
+            MoneyCost = long.Parse(data[idx++]);
             MovingDistance = double.Parse(data[idx++]);
             if (data.Length != 9) throw new ArgumentException("Statistic Data Damaged");
         }
